unified_backend/requirements.txt
# Unified Backend Requirements
# FastAPI and server
fastapi>=0.104.0
uvicorn[standard]>=0.24.0
python-multipart>=0.0.6

# Image processing
pillow>=10.0.0

# AI/ML services
huggingface_hub>=0.19.0
httpx>=0.25.0

# LangGraph and LangChain
langgraph>=0.0.20
langchain>=0.1.0
langchain-community>=0.0.10
langchain-huggingface>=0.0.1

# Data validation
pydantic>=2.0.0

# Configuration
python-dotenv>=1.0.0

# HTTP requests
requests>=2.31.0
